backend/requirements.txt
fastapi
uvicorn
openai
python-multipart
python-pptx
pymupdf
supabase
python-dotenv
pydantic
